+++ test/common/config.js
@@ -53,6 +53,12 @@ export default function () {
         `--plugin-path=${path.join(__dirname, 'fixtures', 'plugins', 'newsfeed')}`,
         `--newsfeed.service.urlRoot=${servers.kibana.protocol}://${servers.kibana.hostname}:${servers.kibana.port}`,
         `--newsfeed.service.pathTemplate=/api/_newsfeed-FTS-external-service-simulators/kibana/v{VERSION}.json`,
+        // Log deprecation warnings on a single line so that they aren't cut off on Buildkite logs
+        '--logging.appenders.deprecation.type=console',
+        '--logging.appenders.deprecation.layout.type=json',
+        '--logging.loggers[0].name=elasticsearch.deprecation',
+        '--logging.loggers[0].level=all',
+        '--logging.loggers[0].appenders[0]=deprecation',
       ],
     },
     services,
